requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# RAGAS and evaluation
ragas==0.1.0
langsmith>=0.0.77
langchain==0.1.0
langchain-openai==0.0.2
httpx==0.25.2
langchain-community==0.0.10
langchain-google-genai==0.0.6

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
openpyxl==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Monitoring
structlog==23.2.0
prometheus-client==0.19.0
